requirements.txt
python-telegram-bot==21.5
requests==2.32.3
flask==3.1.0
